fix(chat): fix slider overload and keyboard arrows navigation (Issue #2825)
